+++ log/src/robocorp/log/__init__.py
@@ -35,7 +35,7 @@ from types import TracebackType
 if typing.TYPE_CHECKING:
     from ._robo_logger import _RoboLogger
 
-__version__ = "2.3.0"
+__version__ = "2.4.0"
 version_info = [int(x) for x in __version__.split(".")]
 
 from . import _config